+++ tests/testthat/helper.R
@@ -4,3 +4,7 @@ temp_desc <- function(file = "D2") {
   file.copy(file, tmp)
   tmp
 }
+
+private <- function(x) {
+  x$.__enclos_env__$private
+}